+++ spec/features/staff/users_can_view_an_organisation_as_xml_spec.rb
@@ -18,7 +18,23 @@ RSpec.feature "Users can view an organisation as XML" do
     end
 
     context "when the user is viewing any other organisation show page" do
-      context "whe organisation has projects" do
+      context "when the organisation has programmes which it is the extending organisation of" do
+        scenario "they can see the download xml button for programmes" do
+          fund = create(:fund_activity)
+          another_fund = create(:fund_activity)
+          _programme = create(:programme_activity, parent: fund, extending_organisation: organisation)
+          _anohter_programme = create(:programme_activity, parent: another_fund, extending_organisation: organisation)
+
+          visit organisation_path(organisation)
+
+          expect(page).to have_link t("page_content.organisation.download.programmes.button", fund_title: fund.title),
+            href: organisation_path(organisation, format: :xml, level: :programme, fund_id: fund.id)
+          expect(page).to have_link t("page_content.organisation.download.programmes.button", fund_title: another_fund.title),
+            href: organisation_path(organisation, format: :xml, level: :programme, fund_id: another_fund.id)
+        end
+      end
+
+      context "when the organisation has projects" do
         scenario "they can download the organisation's projects as XML" do
           _project = create(:project_activity, organisation: organisation)
 
@@ -173,6 +189,17 @@ RSpec.feature "Users can view an organisation as XML" do
 
         expect(page).to have_content(organisation.name)
         expect(page).to_not have_content(t("default.button.download_as_xml"))
+      end
+
+      scenario "they do not see the programmes download buttons" do
+        programme = create(:programme_activity, extending_organisation: organisation)
+        _project = create(:project_activity, parent: programme, organisation: organisation)
+        fund = programme.parent
+
+        visit organisation_path(organisation)
+
+        expect(page).not_to have_link t("page_content.organisation.download.programmes.button", fund_title: fund.title),
+          href: organisation_path(organisation, format: :xml, level: :programme, fund_id: fund.id)
       end
     end